access restriction adn sharing

+++ GardenPlannerApp/Controllers/GardenTilesController.cs
@@ -16,36 +16,30 @@ namespace GardenPlannerApp.Controllers
     [Authorize]
     [Route("api/[controller]")]
     [ApiController]
-    public class GardenTilesController : ControllerBase
+    public class GardenTilesController : GardenPlannerAppControllerBase
     {
-        private readonly GardenPlannerAppDbContext _context;
-        private readonly IMapper _mapper;
 
         public GardenTilesController(GardenPlannerAppDbContext context, IMapper mapper)
-        {
-            _context = context;
-            _mapper = mapper;
-        }
-
+            : base(context, mapper) { }
         // GET: api/GardenTiles
         [HttpGet]
         public async Task<ActionResult<IEnumerable<GardenTileDTO>>> GetGardenTiles()
         {
-            return _mapper.Map<List<GardenTileDTO>>(await _context.GardenTiles.ToListAsync());
+            return SetReadonly(_mapper.Map<List<GardenTileDTO>>(await _context.GardenTiles.OwnedOrPublic(UserId).ToListAsync()));
         }
 
         // GET: api/GardenTiles/5
         [HttpGet("{id}")]
         public async Task<ActionResult<GardenTileDTO>> GetGardenTile(string id)
         {
-            var gardenTile = await _context.GardenTiles.FindAsync(id);
+            var gardenTile = await _context.GardenTiles.OwnedOrPublic(UserId).Where(x => x.Id == id).FirstOrDefaultAsync();
 
             if (gardenTile == null)
             {
                 return NotFound();
             }
 
-            return _mapper.Map<GardenTileDTO>(gardenTile);
+            return SetReadonly(_mapper.Map<GardenTileDTO>(gardenTile));
         }
 
         // PUT: api/GardenTiles/5
@@ -59,7 +53,10 @@ namespace GardenPlannerApp.Controllers
             {
                 return BadRequest();
             }
-
+            if (!Owned(gardenTile))
+            {
+                return Unauthorized();
+            }
             _context.Entry(gardenTile).State = EntityState.Modified;
 
             try
@@ -91,33 +88,38 @@ namespace GardenPlannerApp.Controllers
             {
                 X = newGardenTile.X,
                 Y = newGardenTile.Y,
+                Public = false,
                 Garden = await _context.Gardens.FindAsync(newGardenTile.GardenId),
                 TileType = await _context.TileTypes.FindAsync(newGardenTile.TileTypeId)
             };
+            OwnIt(gardenTile);
             _context.GardenTiles.Add(gardenTile);
             await _context.SaveChangesAsync();
-            return CreatedAtAction("GetGardenTile", new { id = gardenTile.Id }, _mapper.Map<GardenTileDTO>(gardenTile));
+            return CreatedAtAction("GetGardenTile", new { id = gardenTile.Id }, SetReadonly(_mapper.Map<GardenTileDTO>(gardenTile)));
         }
 
         // DELETE: api/GardenTiles/5
         [HttpDelete("{id}")]
         public async Task<ActionResult<GardenTileDTO>> DeleteGardenTile(string id)
         {
-            var gardenTile = await _context.GardenTiles.FindAsync(id);
+            var gardenTile = await _context.GardenTiles.Include(x => x.Owner).Where(x => x.Id == id).FirstOrDefaultAsync();
             if (gardenTile == null)
             {
                 return NotFound();
             }
-
+            if (!Owned(gardenTile))
+            {
+                return Unauthorized();
+            }
             _context.GardenTiles.Remove(gardenTile);
             await _context.SaveChangesAsync();
 
-            return _mapper.Map<GardenTileDTO>(gardenTile);
+            return SetReadonly(_mapper.Map<GardenTileDTO>(gardenTile));
         }
 
         private bool GardenTileExists(string id)
         {
-            return _context.GardenTiles.Any(e => e.Id == id);
+            return _context.GardenTiles.OwnedOrPublic(UserId).Any(e => e.Id == id);
         }
     }
 }
